moved auth folder inside of api folder
