+++ api/auth/auth-router.js
@@ -1,7 +1,7 @@
 const authRouter = require('express').Router();
 const bcrypt = require('bcryptjs');
-const Users = require('../models/users-models');
-const generateToken = require('../middleware/generateToken.js');
+const Users = require('../../models/users-models');
+const generateToken = require('../../middleware/generateToken.js');
 const jwt = require('jsonwebtoken');
 
 module.exports = authRouter;
