+++ api/auth/auth-router.test.js
@@ -1,7 +1,7 @@
 const request = require('supertest');
 const jwt = require('jsonwebtoken');
-const server = require('../api/server');
-const db = require('../database/dbConfig');
+const server = require('../server');
+const db = require('../../database/dbConfig');
 
 beforeEach(async () => {
   await db.from('users').truncate();
